+++ static/js/details.js
@@ -38,7 +38,7 @@ $(function () {
         return false;
     });
 
-    $("#log-status-text").on("click", "#resume-btn", function() {
+    $("#current-status-text").on("click", "#resume-btn", function() {
         $("#resume-form").submit();
         return false;
     });
@@ -77,6 +77,7 @@ $(function () {
     var statusUrl = document.getElementById("events").dataset.statusUrl;
     var lastStatusText = "";
     var lastUpdated = "";
+    var lastStarted = false;
     adaptiveSetInterval(function() {
         $.ajax({
             url: statusUrl + (lastUpdated ? "?u=" + lastUpdated : ""),
@@ -85,10 +86,15 @@ $(function () {
             success: function(data) {
                 if (data.status_text != lastStatusText) {
                     lastStatusText = data.status_text;
-                    $("#log-status-icon").attr("class", "status ic-" + data.status);
-                    $("#log-status-text").html(data.status_text);
+                    $("#current-status-icon").attr("class", "status ic-" + data.status);
+                    $("#current-status-text").html(data.status_text);
 
                     $('#pause-btn').prop('disabled', data.status == "paused");
+                }
+
+                if (data.started != lastStarted) {
+                    lastStarted = data.started;
+                    $("#current-status-spinner").toggleClass("started", data.started);
                 }
 
                 if (data.events) {